backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database and ORM
sqlalchemy==2.0.36
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7

# Swiss Ephemeris for astronomical calculations
pyswisseph==2.10.3.2

# LLM integration
openai==2.5.0
httpx==0.27.2

# Caching
redis==5.0.1

# Validation and serialization
pydantic==2.8.2
pydantic-settings==2.2.1
email-validator==2.2.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Rate limiting
slowapi==0.1.9

# Compression
gzip-stream==1.2.0

